+++ HISTORY.rst
@@ -2,6 +2,11 @@
 
 History
 
+0.2.0 (2016-08-07)
+-------------------
+* Even cache when got empty results
+* When exception, use old data to cache degraded
+
 0.1.7 (2016-4-14)
 -----------------
 * cache refresh manager, should be used with lock